# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Legacy Sports API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/lookup-school:
    post:
      operationId: lookupSchool
      tags: [auth]
      summary: Look up school by code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolCodeInput"
      responses:
        "200":
          description: School found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolPublic"
        "404":
          description: School not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register player or coach
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /performances:
    get:
      operationId: listPerformances
      tags: [performances]
      summary: List performances
      parameters:
        - name: playerId
          in: query
          schema:
            type: integer
        - name: coachId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Performance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Performance"
    post:
      operationId: createPerformance
      tags: [performances]
      summary: Log a performance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerformanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Performance"
  /performances/{id}:
    delete:
      operationId: deletePerformance
      tags: [performances]
      summary: Delete performance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: schoolId
          in: query
          schema:
            type: integer
        - name: coachId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: markAttendance
      tags: [attendance]
      summary: Mark attendance for players
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceBatch"
      responses:
        "201":
          description: Attendance marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    SchoolCodeInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
    SchoolPublic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        logoUrl:
          type: string
        primaryColor:
          type: string
        isPaused:
          type: boolean
        pauseMessage:
          type: string
      required: [id, name, code]
...
    LoginInput:
      type: object
      required: [schoolCode, username, password]
      properties:
        schoolCode:
          type: string
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/SessionUser"
        school:
          $ref: "#/components/schemas/SchoolPublic"
      required: [user]
    SessionUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        role:
          type: string
        schoolId:
          type: ["integer", "null"]
        sport:
          type: string
        status:
          type: string
        isOwner:
          type: boolean
      required: [id, name, username, role]
    RegistrationInput:
      type: object
      required: [schoolCode, role, username, password, name, sport]
      properties:
        schoolCode:
          type: string
        role:
          type: string
          enum: [player, coach]
        username:
          type: string
        password:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsappNumber:
          type: string
        address:
          type: string
        sport:
          type: string
        admissionNumber:
          type: string
        age:
          type: integer
        dateOfBirth:
          type: string
        gender:
          type: string
        className:
          type: string
        section:
          type: string
        rollNumber:
          type: string
        designation:
          type: string
        parentName:
          type: string
        parentPhone:
          type: string
        parentWhatsapp:
          type: string
        parentEmail:
          type: string
...
    Performance:
      type: object
      properties:
        id:
          type: integer
        playerId:
          type: integer
        coachId:
          type: ["integer", "null"]
        sport:
          type: string
        sessionType:
          type: string
        sessionNotes:
          type: string
        customData:
          type: object
          additionalProperties: true
        recordedAt:
          type: string
        createdAt:
          type: string
      required: [id, playerId, sport]
    PerformanceInput:
      type: object
      required: [playerId, sport, customData]
      properties:
        playerId:
          type: integer
        sport:
          type: string
        sessionType:
          type: string
        sessionNotes:
          type: string
        customData:
          type: object
          additionalProperties: true
        recordedAt:
          type: string
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        schoolId:
          type: integer
        coachId:
          type: ["integer", "null"]
        attDate:
          type: string
        status:
          type: string
        sessionType:
          type: string
        notes:
          type: string
        createdAt:
          type: string
        userName:
          type: string
      required: [id, userId, schoolId, attDate, status]
    AttendanceMark:
      type: object
      required: [userId, status]
      properties:
        userId:
          type: integer
        status:
          type: string
        notes:
          type: string
    AttendanceBatch:
      type: object
      required: [schoolId, attDate, records]
      properties:
        schoolId:
          type: integer
        attDate:
          type: string
        sessionType:
          type: string
        records:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceMark"
...